tests/wpt/web-platform-tests/bluetooth/server/disconnect/gc-detach.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Garbage collect then detach frame. We shouldn\'t crash.';
let iframe = document.createElement('iframe');

bluetooth_test(() => setUpConnectableHealthThermometerDevice()
    // 1. Load the iframe.
    .then((f) => new Promise(resolve => {
      iframe.src = '/bluetooth/resources/health-thermometer-iframe.html';
      document.body.appendChild(iframe);
      iframe.addEventListener('load', resolve);
    }))
    // 2. Connect device, run garbage collection, and detach iframe.
    .then(() => new Promise(resolve => {
      callWithTrustedClick(() => {
        iframe.contentWindow.postMessage({
          type: 'RequestAndConnect',
          options: {filters: [{services: ['health_thermometer']}]}
        }, '*');
      });

      window.onmessage = messageEvent => {
        assert_equals(messageEvent.data, 'Connected');
        runGarbageCollection().then(() => {
          iframe.remove();
          resolve();
        });
      }
    })), test_desc)
</script>
</body>
